integrate Bosses with Mob system

+++ src/main/java/me/redplayer_1/custombosses/entity/Mob.java
@@ -3,16 +3,15 @@ package me.redplayer_1.custombosses.entity;
 import eu.decentsoftware.holograms.api.DHAPI;
 import eu.decentsoftware.holograms.api.holograms.Hologram;
 import me.redplayer_1.custombosses.CustomBosses;
+import me.redplayer_1.custombosses.events.DamageListener;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
-import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.LivingEntity;
-import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.HandlerList;
 import org.bukkit.event.Listener;
-import org.bukkit.event.entity.EntityDamageByEntityEvent;
+import org.bukkit.event.entity.EntityDamageEvent;
 import org.bukkit.metadata.FixedMetadataValue;
 import org.bukkit.metadata.MetadataValue;
 import org.jetbrains.annotations.Nullable;
@@ -28,22 +27,39 @@ public class Mob implements Listener {
     private final UUID uuid = UUID.randomUUID();
 
     // incoming damage is multiplied by this. Affected by custom armor & weapons (+ enchants)
-    private double naturalDamageScalar = 1.0;
-    private double playerDamageScalar = 1.0;
+    private double damageScalar = 1.0;
 
     private LivingEntity entity;
+    private boolean invincible;
     private final double maxHealth;
     private double health; // the actual health of this mob
-    private Hologram hologram; // line 0 is the name, line 1 is the health
-    /*
-    About:
-    - health (independent of entity)
-    - type (bukkit mob/entity)
-    - name (as holo = color support)
-    - use paper goal api
-     */
 
-    // TODO: name may contain color codes?
+    private Hologram hologram; // line 0 is the name, line 1 is the health
+
+    /**
+     * Create a new {@link Mob} that is automatically spawned.
+     * @param name the display name (can contain color codes)
+     * @param type type of mob to spawn (Must be a {@link LivingEntity})
+     * @param location the Location to spawn the Mob at
+     * @param maxHealth the maximum health
+     * @param attackRange how far away entities can be attacked form (distance in blocks ^ 2)
+     * @param hostile whether the Mob should attack other entities
+     * @param invincible if the Mob should take damage
+     */
+    public Mob(String name, EntityType type, Location location, double maxHealth, int attackRange, boolean hostile, boolean invincible) {
+        this(name, type, location, maxHealth, attackRange, hostile);
+        this.invincible = invincible;
+    }
+
+    /**
+     * Create a new {@link Mob} that is automatically spawned.
+     * @param name the display name (can contain color codes)
+     * @param type type of mob to spawn (Must be a {@link LivingEntity})
+     * @param location the Location to spawn the Mob at
+     * @param maxHealth the maximum health
+     * @param attackRange how far away entities can be attacked form (distance in blocks ^ 2)
+     * @param hostile whether the Mob should attack other entities
+     */
     public Mob(String name, EntityType type, Location location, double maxHealth, int attackRange, boolean hostile) {
         // ensure entity is a living one
         if (location.getWorld().spawnEntity(location, type) instanceof LivingEntity le) {
@@ -92,19 +108,39 @@ public class Mob implements Listener {
      * @param amount amount to damage (hit points)
      */
     public void damage(double amount) {
-        health -= amount;
-        if (health <= 0) {
+        if (health - amount <= 0) {
             kill();
+        } else {
+            health -= amount;
         }
     }
 
     /**
      * Kills the Mob and its corresponding Entity
+     * @return if the Mob was successfully killed
      */
-    public void kill() {
-        HandlerList.unregisterAll(this);
-        entity.setHealth(0);
-        hologram.delete();
+    public boolean kill() {
+        return !kill(false);
+    }
+
+    /**
+     * Kills the Mob and its corresponding Entity
+     * @param force if the Mob should die regardless of whether the {@link MobDeathEvent} was cancelled
+     * @return if the event was cancelled
+     */
+    public boolean kill(boolean force) {
+        double prevHealth = health;
+        health = 0; // make isDead() return true
+        MobDeathEvent event = new MobDeathEvent(this, DamageListener.getLastDamager(uuid));
+        event.callEvent();
+        if (force || !event.isCancelled()) {
+            HandlerList.unregisterAll(this);
+            entity.setHealth(0);
+            hologram.delete();
+        } else {
+            health = prevHealth;
+        }
+        return event.isCancelled();
     }
 
     public boolean isDead() {
@@ -119,19 +155,60 @@ public class Mob implements Listener {
         return entity.getLocation();
     }
 
+    /**
+     * @return the Bukkit representation of the entity (health changes not reflected, @see )
+     */
     public LivingEntity getEntity() {
         return entity;
     }
 
+    /**
+     * Set the health of the Mob. Values greater than the max will be rounded down.
+     */
+    public void setHealth(double health) {
+        setHealth(health, true);
+    }
+
+    /**
+     * Sets the health of the Mob. Values greater than the max can be used by setting ignoreMax to true.
+     * @param ignoreMax whether values greater than the maximum should be allowed
+     */
+    public void setHealth(double health, boolean ignoreMax) {
+        if (ignoreMax) this.health = health;
+        else this.health = Math.min(health, maxHealth);
+    }
+
+    public void setInvincible(boolean invincible) {
+        this.invincible = invincible;
+    }
+
+    /**
+     * Sets {@link Mob#invincible} to the value specified for the duration, then reverts it back.
+     * @param invincible value to set
+     * @param ticks how long to set it for (in ticks)
+     */
+    public void setInvincible(boolean invincible, long ticks) {
+        boolean previous = this.invincible;
+        this.invincible = invincible;
+        Bukkit.getScheduler().runTaskLater(CustomBosses.getInstance(), () -> setInvincible(previous), ticks);
+    }
+
+    /**
+     * Makes the Mob invincible for the duration and then makes it vincible again.
+     * @param ticks how long the Mob should be invincible (in ticks)
+     */
+    public void setInvincible(long ticks) {
+        invincible = true;
+        Bukkit.getScheduler().runTaskLater(CustomBosses.getInstance(), () -> invincible = false, ticks);
+    }
+
     @EventHandler
-    public void onEntityDamageByEntity(EntityDamageByEntityEvent event) {
+    public void onEntityDamage(EntityDamageEvent event) {
         if (event.getEntity().equals(entity)) {
             // damage is affected by scalars, armor, and weapons
-            if (event.getDamager() instanceof Player) {
-                damage(event.getFinalDamage() * playerDamageScalar);
-            } else {
-                damage(event.getFinalDamage() * naturalDamageScalar);
-            }
+            if (!invincible)
+                damage(event.getFinalDamage() * damageScalar);
+            else event.setCancelled(true);
             event.setDamage(0);
         }
     }
